requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
google-generativeai==0.3.0
PyJWT==2.3.0
bcrypt==3.2.0
Werkzeug==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.9
Faker==22.6.0 